+++ lib/screens/authentications/home_screen.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:voicepocket/constants/gaps.dart';
 import 'package:voicepocket/constants/sizes.dart';
-import 'package:voicepocket/screens/recordroom_screen.dart';
-import 'package:voicepocket/screens/voicepocket_play_screen.dart';
+import 'package:voicepocket/screens/recordroom/recordroom_screen.dart';
+import 'package:voicepocket/screens/voicepocket/voicepocket_play_screen.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
@@ -19,7 +19,7 @@ class HomeScreen extends StatelessWidget {
   void _onVoiceTab(BuildContext context) {
     Navigator.of(context).push(
       MaterialPageRoute(
-        builder: (context) => VoicePocketPlayScreen(),
+        builder: (context) => const VoicePocketPlayScreen(),
       ),
     );
   }
