+++ lib/screens/voicepocket/media_player_screen.dart
@@ -1,0 +1,148 @@
+import 'dart:io';
+
+import 'package:audioplayers/audioplayers.dart';
+import 'package:flutter/material.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:voicepocket/constants/gaps.dart';
+import 'package:voicepocket/constants/sizes.dart';
+
+class MediaPlayerScreen extends StatefulWidget {
+  final String path;
+  const MediaPlayerScreen({super.key, required this.path});
+
+  @override
+  State<MediaPlayerScreen> createState() => _MediaPlayerScreenState();
+}
+
+class _MediaPlayerScreenState extends State<MediaPlayerScreen> {
+  final audioPlayer = AudioPlayer();
+  bool isPlaying = false;
+  Duration duration = Duration.zero;
+  Duration position = Duration.zero;
+
+  @override
+  void initState() {
+    super.initState();
+    setAudio();
+    audioPlayer.onPlayerStateChanged.listen((state) {
+      setState(() {
+        isPlaying = state == PlayerState.playing;
+      });
+    });
+    audioPlayer.onDurationChanged.listen((newDuration) {
+      setState(() {
+        duration = newDuration;
+      });
+    });
+    audioPlayer.onPositionChanged.listen((newPosition) {
+      setState(() {
+        position = newPosition;
+      });
+    });
+  }
+
+  Future setAudio() async {
+    audioPlayer.setReleaseMode(ReleaseMode.loop);
+
+    var directory = await getApplicationDocumentsDirectory();
+    final file = File("${directory.path}/${widget.path}");
+    audioPlayer.setSourceDeviceFile(file.path);
+  }
+
+  String formatTime(Duration duration) {
+    String twoDigits(int n) => n.toString().padLeft(2, "0");
+    final hours = twoDigits(duration.inHours);
+    final minutes = twoDigits(duration.inMinutes);
+    final seconds = twoDigits(duration.inSeconds);
+
+    return [
+      if (duration.inHours > 0) hours,
+      minutes,
+      seconds,
+    ].join(":");
+  }
+
+  @override
+  void dispose() {
+    audioPlayer.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: Padding(
+        padding: const EdgeInsets.all(20),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            ClipRRect(
+              borderRadius: BorderRadius.circular(20),
+              child: Image.network(
+                "https://developer.android.com/static/images/jetpack/compose/graphics-sourceimageland.jpg?hl=ko",
+                width: double.infinity,
+                height: 350,
+                fit: BoxFit.cover,
+              ),
+            ),
+            Gaps.v32,
+            const Text(
+              "The Flutter Song",
+              style: TextStyle(
+                fontSize: Sizes.size24,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+            Gaps.v3,
+            const Text(
+              "Sarah Abs",
+              style: TextStyle(
+                fontSize: Sizes.size20,
+              ),
+            ),
+            Slider(
+              min: 0,
+              max: duration.inSeconds.toDouble(),
+              value: position.inSeconds.toDouble(),
+              onChanged: (value) async {
+                final position = Duration(seconds: value.toInt());
+                await audioPlayer.seek(position);
+
+                await audioPlayer.resume();
+              },
+            ),
+            Padding(
+              padding: const EdgeInsets.symmetric(
+                horizontal: Sizes.size16,
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Text(formatTime(position)),
+                  Text(formatTime(duration)),
+                ],
+              ),
+            ),
+            CircleAvatar(
+              radius: 35,
+              child: IconButton(
+                onPressed: () async {
+                  if (isPlaying) {
+                    await audioPlayer.pause();
+                  } else {
+                    await audioPlayer.resume();
+                  }
+                },
+                icon: Icon(
+                  isPlaying ? Icons.pause : Icons.play_arrow,
+                ),
+                iconSize: Sizes.size52,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
