+++ lib/screens/voicepocket/voicepocket_list_screen.dart
@@ -1,30 +1,32 @@
 import 'package:flutter/material.dart';
 
-void main() => runApp(MyApp());
-
-
+void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Appbar',
       theme: ThemeData(primarySwatch: Colors.red),
-      home: MyPage(),
+      home: const MyPage(),
     );
   }
 }
 
 class MyPage extends StatelessWidget {
+  const MyPage({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Appbar Icon Menu'),
+        title: const Text('Appbar Icon Menu'),
         centerTitle: true,
         elevation: 0.0,
         leading: IconButton(
-          icon: Icon(Icons.menu),
+          icon: const Icon(Icons.menu),
           onPressed: () {
             print('menu button is clicked');
           },
@@ -32,7 +34,7 @@ class MyPage extends StatelessWidget {
         actions: [
           //이곳에 한개 이상의 위젯들을 가진다.
           IconButton(
-            icon: Icon(Icons.shopping_cart),
+            icon: const Icon(Icons.shopping_cart),
             onPressed: () {
               print('search button is clicked');
             },
