+++ lib/services/stream_to_audio.dart
@@ -1,0 +1,20 @@
+import 'dart:typed_data';
+import 'package:just_audio/just_audio.dart';
+
+class StreamToAudio extends StreamAudioSource {
+  final Uint8List _buffer;
+
+  StreamToAudio(this._buffer) : super(tag: 'MyAudioSource');
+
+  @override
+  Future<StreamAudioResponse> request([int? start, int? end]) async {
+    // Returning the stream audio response with the parameters
+    return StreamAudioResponse(
+      sourceLength: _buffer.length,
+      contentLength: (start ?? 0) - (end ?? _buffer.length),
+      offset: start ?? 0,
+      stream: Stream.fromIterable([_buffer.sublist(start ?? 0, end)]),
+      contentType: 'audio/wav',
+    );
+  }
+}
